requirements.txt
Flask==2.3.3
Markdown==3.5.1
PyYAML==6.0.1
Pygments==2.16.1
python-dateutil==2.8.2
Pillow>=10.1
pillow-heif==0.13.0
